pyproject.toml
[tool.poetry]
name = "many2one-route-optimization"
version = "0.1.0"
description = "Optimize many-to-one routes using Google Maps API"
authors = ["Stas German <stas.german@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/StGerman/many2one-route-optimization"
repository = "https://github.com/StGerman/many2one-route-optimization"
documentation = "https://github.com/StGerman/many2one-route-optimization/docs"
keywords = ["routing", "optimization", "many-to-one", "google-api"]

classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28"
pandas = "^1.5"
pyyaml = "^6.0"
python-dotenv = "^0.21"
scipy = "^1.10"

[tool.poetry.dev-dependencies]
pytest = "^7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
